osvm::services

Module account_decoders

Account data decoders for known Solana programs

This module provides decoders for various Solana program account types, allowing the snapshot reader to display human-readable decoded information.

Structs

AddressLookupTableData

AddressLookupTableDecoder

Address Lookup Table Decoder

BpfUpgradeableDecoder

BPF Upgradeable Loader Decoder

BpfUpgradeableProgramData

CandyMachineData

CandyMachineDecoder

Candy Machine Decoder

CreatorData

DecoderRegistry

Decoder registry for managing all decoders

DriftV2Data

FeeCalculatorData

GovernanceData

GovernanceDecoder

SPL Governance Decoder

JupiterAggregatorData

JupiterLimitOrderData

KaminoLendingData

LidoSolanaData

LifinityAmmData

MagicEdenV2Data

MangoV4Data

MarinadeFinanceData

MemoData

MemoDecoder

Memo Program Decoder

MetaplexBubblegumData

MetaplexMetadataDecoder

Metaplex Token Metadata Decoder

MeteoraPoolsData

NameServiceData

NameServiceDecoder

Name Service Decoder

NonceAccountData

NonceDecoder

System Program Nonce Decoder

OpenbookV2Data

OrcaWhirlpoolData

OrcaWhirlpoolDecoder

Orca Whirlpool Decoder

PhoenixV1Data

PythOracleData

PythOracleDecoder

Pyth Oracle Decoder

RaydiumAmmData

RaydiumAmmDecoder

Raydium AMM Decoder

SerumMarketData

SerumMarketDecoder

Serum Market Decoder

SplStakePoolData

SplTokenDecoder

SPL Token Account Decoder

SplTokenMintDecoder

SPL Token Mint Decoder

SquadsV3Data

StakeAccountData

StakeDecoder

Stake Account Decoder

SwitchboardData

SwitchboardDecoder

Switchboard Decoder

TensorSwapData

Token2022Data

TokenAccountData

TokenMetadataData

TokenMintData

VoteAccountData

VoteDecoder

Vote Account Decoder

Enums

DecodedAccount

Decoded account data variants

Constants

ADDRESS_LOOKUP_TABLE_ID

ADRENA_ID

ALDRIN_SWAP_ID

ALLBRIDGE_ID

APRICOT_FINANCE_ID

AURORY_ID

BPF_LOADER_UPGRADEABLE_ID

CANDY_MACHINE_V2_ID

CONFIG_PROGRAM_ID

CROPPER_ID

CYPHER_PROTOCOL_ID

DRIFT_V2_ID

ENTROPY_ID

FLASH_PROTOCOL_ID

FLASH_TRADE_ID

FRANCIUM_ID

FRIKTION_ID

GENOPETS_ID

HAWKSIGHT_ID

HELIUM_ID

HXRO_NETWORK_ID

JET_PROTOCOL_ID

JUPITER_AGGREGATOR_V6_ID

JUPITER_LIMIT_ORDER_ID

KAMINO_LENDING_ID

KATANA_ID

LARIX_ID

LIDO_SOLANA_ID

LIFINITY_AMM_ID

MAGIC_EDEN_V2_ID

MANGO_V4_ID

MARINADE_FINANCE_ID

MEMO_PROGRAM_ID

MEMO_PROGRAM_V1_ID

MERCURIAL_SWAP_ID

METAPLEX_BUBBLEGUM_ID

METAPLEX_TOKEN_METADATA_ID

METEORA_POOLS_ID

NAME_SERVICE_ID

O1_EXCHANGE_ID

OPENBOOK_V2_ID

ORCA_WHIRLPOOL_ID

PARCL_ID

PHOENIX_V1_ID

PORTAL_BRIDGE_ID

PORT_FINANCE_ID

PSY_OPTIONS_ID

PYTH_ORACLE_PROGRAM_ID

QUARRY_MINE_ID

RAGE_TRADE_ID

RAYDIUM_AMM_V4_ID

RENDER_NETWORK_ID

RIBBON_FINANCE_ID

SABER_SWAP_ID

SERUM_DEX_V3_ID

SOLEND_ID

SPL_ASSOCIATED_TOKEN_ACCOUNT_PROGRAM_ID

SPL_GOVERNANCE_ID

SPL_STAKE_POOL_ID

SPL_TOKEN_2022_ID

SPL_TOKEN_PROGRAM_ID

SPREADS_ID

SQUADS_V3_ID

STAKE_PROGRAM_ID

STAR_ATLAS_ID

STEPN_ID

SWITCHBOARD_PROGRAM_ID

SYMMETRY_ID

SYSTEM_PROGRAM_ID

TENSOR_SWAP_ID

THETANUTS_ID

TULIP_PROTOCOL_ID

VOTE_PROGRAM_ID

WORMHOLE_CORE_ID

ZETA_MARKETS_ID

Traits

AccountDecoder

Main decoder trait

Functions

format_decoded_account

Helper function to format decoded account for display