osvm::utils

Module local_rpc

Local RPC node deployment and management Local RPC node deployment and management

This module provides functionality to deploy and manage RPC nodes on localhost for development and testing purposes.

Structs

LocalRpcConfig

Configuration for local RPC node deployment

LocalRpcInfo

Information about a running local RPC node

LocalRpcStatus

Status information for local RPC node

Functions

check_local_rpc_status

Check the status of local RPC node

start_local_rpc

Start a local RPC node

stop_local_rpc

Stop the local RPC node