osvm::utils

Module mainnet_rpc

Mainnet RPC connectivity Mainnet RPC node deployment and management

This module provides functionality to deploy and manage a mainnet RPC node that syncs with the real Solana mainnet-beta blockchain.

Structs

MainnetRpcConfig

Configuration for mainnet RPC node deployment

MainnetRpcInfo

Information about a running mainnet RPC node

MainnetRpcStatus

Status information for mainnet RPC node

Constants

MAINNET_ENTRYPOINTS 🔒

Known mainnet entrypoints

Functions

check_mainnet_rpc_status

Check the status of mainnet RPC node

start_mainnet_rpc

Start a mainnet RPC node

stop_mainnet_rpc

Stop the mainnet RPC node